NodeMCU 1
The code is to receive data from the dht 11 sensor which measures temperature and humidity along with the soil moisture sensor that detects the amount of moisture in the soil. The sensors are connected through a NodeMCU to use the wifi features in order to integrate the values from the sensors to the blynk ioT application.