NodeMCU 2
The file consists of the Arduino C code to connect the DC motors in order to pump water. The code uses a DC motor that is integrated into Blynk ioT app in which case the flow of water can be controlled by the user through the slider widget. The NodeMCU is connected to the DC motor to supply power and transmit the inputs from the blynk app through wifi.