NodeMCU 1
The file consists of arduino C code that connects various sensors to the nodemcu in order to integrate the values into the bynk ioT app. The various sensors in this circuitry and code include the soil moisture sensor which detects the amount of moisture in the soil along with the Dht11 which measures the temperature and humidity of the surroundings. The nodemcu is used in this circuitry to provide power and communication over wlan, or wireless communication. The blynk functions are added in order to transfer the values from the sensors to blynk with the use of wifi