Arduino 1
The file consists of Arduino C code that is particularly used to connect sensors to the microcontroller in order to transmit the data through the GSM module. The code uses sensors such as LDR and smoke sensors in order to measure the amount of light received and the quality of air. The user is immediately informed through an SMS message which is sent if the environment crops aren't ideal. The code uses the AT commands to send the data through the GSM module which contains a slot for a SIM card to send the message to the desired mobile number mentioned in the code.